chore: Add page.c / page.h (source from vm)
