+++ vm/targets.mk
@@ -3,3 +3,4 @@ vm_SRC += vm/uninit.c     # Uninitialized page
 vm_SRC += vm/anon.c       # Anonymous page
 vm_SRC += vm/file.c       # File mapped page
 vm_SRC += vm/inspect.c    # Testing utility
+vm_SRC += vm/page.c 			# Supplementary Page Table
